bot/requirements.txt
python-telegram-bot
psycopg2-binary
python-dotenv
watchdog
uvicorn
fastapi
sqlalchemy